--- src/main/java/duke/GoldBot.java
@@ -1,35 +1,0 @@
-package duke;
-import java.io.IOException;
-import java.util.Scanner;
-
-import duke.exceptions.FileCorruptionException;
-
-/**
- * GoldBot class to handle main bot operations
- */
-public class GoldBot {
-    public static void main(String[] args) {
-        System.out.println("test");
-        Storage storage = new Storage("./data/data.txt");
-        TaskList taskList;
-        try {
-            taskList = storage.createTaskList();
-        } catch (IOException e) {
-            System.out.println("Unexpected error occured.");
-            return;
-        } catch (FileCorruptionException e) {
-            System.out.println("File corrupted.");
-            return;
-        }
-        ChatSession session = new ChatSession(taskList);
-        session.initChat();
-        
-        Scanner scanner = new Scanner(System.in);
-
-        while (session.shouldContinueSession) {
-            String input = scanner.nextLine();
-            session.handleMessage(input);
-        }
-        scanner.close();
-    }
-}
