+++ src/main/java/duke/Launcher.java
@@ -3,6 +3,6 @@ import javafx.application.Application;
 
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(GoldBotGui.class, args);
+        Application.launch(Ui.class, args);
     }
 }
