+++ src/main/java/duke/Ui.java
@@ -1,28 +1,131 @@
 package duke;
-public class Ui {
+import java.io.IOException;
+
+import duke.exceptions.FileCorruptionException;
+import javafx.application.Application;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
+import javafx.scene.layout.Region;
+import javafx.scene.control.Label;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+
+
+public class Ui extends Application {
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+    private Storage storage;
+    private TaskList taskList;
     private ChatSession session;
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
 
-    Ui(ChatSession session) {
-        this.session = session;
+    public void start(Stage stage) {
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
+        
+        userInput = new TextField();
+        sendButton = new Button("Send");
+
+        AnchorPane mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+
+        stage.setTitle("Duke");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+
+        mainLayout.setPrefSize(400.0, 600.0);
+
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        //You will need to import `javafx.scene.layout.Region` for this.
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(325.0);
+
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        Scene scene = new Scene(mainLayout, 400, 200);
+        stage.setScene(scene);
+        stage.setTitle("GoldBot");
+        stage.show();
+
+        Storage storage = new Storage("./data/data.txt");
+        TaskList taskList;
+        try {
+            taskList = storage.createTaskList();
+        } catch (IOException e) {
+            printDukeMessage("Unexpected error occured.");
+            return;
+        } catch (FileCorruptionException e) {
+            printDukeMessage("File corrupted.");
+            return;
+        }
+        session = new ChatSession(taskList, this);
+
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput();
+        });
+    
+        userInput.setOnAction((event) -> {
+            handleUserInput();
+        });
+
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
     }
 
-    public void initChat() {
-        String message = "____________________________________________________________\n" +
-        "Hello! I'm GoldBot!\n" +
-        "What can I do for you?\n" +
-        "____________________________________________________________\n";
-        System.out.println(message);
+    private Label getDialogLabel(String text) {
+    // You will need to import `javafx.scene.control.Label`.
+        Label textToAdd = new Label(text);
+        textToAdd.setWrapText(true);
+
+        return textToAdd;
     }
 
-    public void printMessage(String message) {
-        this.printHorizontalLine();
-        System.out.println(message);
-        this.printHorizontalLine();
+    public void printDukeMessage(String message) {
+        dialogContainer.getChildren().add(
+            DialogBox.getDukeDialog(getDialogLabel(message), new ImageView(duke))
+        );
     }
 
-    public void printHorizontalLine() {
-        System.out.println("____________________________________________________________");
+    public void printUserMessage(String message) {
+        dialogContainer.getChildren().add(
+            DialogBox.getUserDialog(getDialogLabel(message), new ImageView(user))
+        );
     }
 
+    private void handleUserInput() {
+        printUserMessage(userInput.getText());
+        getResponse(userInput.getText());
+        userInput.clear();
+    }
 
+    private void getResponse(String input) {
+        session.handleMessage(input);
+    }
 }
